+++ Devices/ConnectionCoordinator.cs
@@ -490,6 +490,7 @@ namespace MozaPlugin.Devices
             _detectionState.BaseAmbientProbed = false;
             _detectionState.BaseEq10Probed = false;
             _detectionState.BaseFwVersionLogged = false;
+            _detectionState.BaseFwVersionProbeRetries = 0;
             _detectionState.BaseOwner = null;
             _data.BaseSettingsRead = false;
             try { _plugin.PendingResponses.Clear(); } catch { }
@@ -790,6 +791,7 @@ namespace MozaPlugin.Devices
             _detectionState.BaseAmbientProbed = false;
             _detectionState.BaseEq10Probed = false;
             _detectionState.BaseFwVersionLogged = false;
+            _detectionState.BaseFwVersionProbeRetries = 0;
             _data.IsBaseConnected = false;
             _data.BaseSettingsRead = false;
             var baseDm = _baseManager?.DeviceManager;